Add CORS rules when enabling Media Services module + AMP style fix
