+++ src/Orchard.Web/Modules/Orchard.Azure.MediaServices/Content/amp/stylefix.css
@@ -10,3 +10,8 @@
 .azuremediaplayer.amp-default-skin > button:hover {
     background-color: rgba(255,255,255,.1);
 }
+
+.azuremediaplayer.amp-default-skin.amp-size-s {
+    /* workaround for AMP style issue on small screens */
+    font-size: 400px !important;
+}
