+++ src/Orchard.Web/Modules/Orchard.Azure.MediaServices/Orchard.Azure.MediaServices.csproj
@@ -260,6 +260,7 @@
     <Content Include="Content\expressInstall.swf" />
     <Content Include="Content\MSAdaptiveStreamingPlugin.swf" />
     <Content Include="Content\StrobeMediaPlayback.swf" />
+    <Compile Include="Helpers\StorageHelper.cs" />
     <Compile Include="Models\EncodingPreset.cs" />
     <Compile Include="Shapes\CloudVideoPlayerShape.cs" />
     <Compile Include="Helpers\BitExtensions.cs" />
